requirements.txt
asgiref==3.8.1
Django==4.2
django-filter==24.2
djangorestframework==3.15.1
Markdown==3.6
pillow==10.3.0
python-decouple==3.8
sqlparse==0.4.4
typing_extensions==4.11.0
